@@ -1,21 +1,30 @@
 {
-  "name":              "jukylin/jaeger-php",
-  "description":       "php client for jaeger",
-  "keywords":          ["jaeger","trace","opentracing"],
-  "license":           "Apache-2.0",
+  "name": "auxmoney/jaeger-php",
+  "description": "php client for jaeger",
+  "keywords": [
+    "jaeger",
+    "trace",
+    "opentracing"
+  ],
+  "license": "Apache-2.0",
   "minimum-stability": "stable",
-  "require":           {
-    "php": ">=5.6.0",
-    "packaged/thrift" : "0.10.0",
-    "opentracing/opentracing" : "1.0.0-beta5"
+  "require": {
+    "php": ">=7.1",
+    "ext-json": "*",
+    "packaged/thrift": "~0.13.0",
+    "opentracing/opentracing": "^1.0.1"
   },
   "authors": [
     {
       "name": "jukylin",
       "email": "742161455@qq.com"
+    },
+    {
+      "name": "auxmoney GmbH",
+      "email": "opensource@auxmoney.com"
     }
   ],
-  "autoload":          {
+  "autoload": {
     "psr-4": {
       "Jaeger\\": "src\\Jaeger"
     },
@@ -23,11 +32,22 @@
       "src/Jaeger/Constants.php"
     ]
   },
+  "autoload-dev": {
+    "psr-4": {
+      "tests\\": "tests\\"
+    }
+  },
   "scripts": {
-    "test": "./vendor/bin/phpunit"
+    "test": "phpdbg -qrr vendor/bin/phpunit",
+    "phpcs": "vendor/bin/php-cs-fixer fix --ansi",
+    "phpcs-check": "vendor/bin/php-cs-fixer fix --dry-run --diff 1>&2",
+    "phpstan": "vendor/bin/phpstan analyse src tests"
   },
   "require-dev": {
-    "phpunit/phpunit": "^5",
-    "php-coveralls/php-coveralls": "^1.0"
+    "roave/security-advisories": "dev-latest",
+    "phpunit/phpunit": "^8.5",
+    "php-coveralls/php-coveralls": "^v2.4",
+    "phpstan/phpstan": "^0.12",
+    "friendsofphp/php-cs-fixer": "^v3.1"
   }
 }
