@@ -1,14 +1,10 @@
 <?php
 
-require_once dirname(dirname(dirname(dirname(__FILE__)))).'/autoload.php';
+require_once dirname(__FILE__, 2).'/vendor/autoload.php';
 
 use Jaeger\Config;
-use GuzzleHttp\Client;
-use OpenTracing\Formats;
-use OpenTracing\Reference;
 
 unset($_SERVER['argv']);
-
 
 //init server span start
 $config = Config::getInstance();
@@ -20,21 +16,20 @@ $second = $tracer->startActiveSpan('level second');
 $third = $tracer->startActiveSpan('level third');
 
 $num = 0;
-for ($i = 0; $i < 10; $i++){
-    $num += 1;
+for ($i = 0; $i < 10; ++$i) {
+    ++$num;
 }
-$third->getSpan()->setTag("num", $num);
+$third->getSpan()->setTag('num', $num);
 sleep(1);
 $third->close();
 
 $num = 0;
-for ($i = 0; $i < 10; $i++){
+for ($i = 0; $i < 10; ++$i) {
     $num += 2;
 }
-$third->getSpan()->setTag("num", $num);
+$third->getSpan()->setTag('num', $num);
 sleep(1);
 $second->close();
-
 
 $top->close();
 
