@@ -15,158 +15,172 @@
 
 namespace Jaeger;
 
-use Jaeger\Reporter\RemoteReporter;
-use Jaeger\Reporter\Reporter;
-use Jaeger\Transport\TransportUdp;
-use OpenTracing\NoopTracer;
-use Jaeger\Sampler\Sampler;
-use Jaeger\Sampler\ConstSampler;
+use const Jaeger\Constants\PROPAGATOR_JAEGER;
+use const Jaeger\Constants\PROPAGATOR_ZIPKIN;
 use Jaeger\Propagator\JaegerPropagator;
 use Jaeger\Propagator\ZipkinPropagator;
+use Jaeger\Reporter\RemoteReporter;
+use Jaeger\Reporter\Reporter;
+use Jaeger\Sampler\ConstSampler;
+use Jaeger\Sampler\Sampler;
+use Jaeger\Transport\Transport;
+use Jaeger\Transport\TransportUdp;
+use OpenTracing\NoopTracer;
+use OpenTracing\Span as OpenTracingSpan;
+use OpenTracing\Tracer;
+use RuntimeException;
 
-class Config {
+class Config
+{
+    /**
+     * @var Transport|null
+     */
+    private $transport;
 
-    private $transport = null;
+    /**
+     * @var Reporter|null
+     */
+    private $reporter;
 
-    private $reporter = null;
+    /**
+     * @var Sampler|null
+     */
+    private $sampler;
 
-    private $sampler = null;
-
-    private $scopeManager = null;
+    /**
+     * @var ScopeManager|null
+     */
+    private $scopeManager;
 
     private $gen128bit = false;
 
-    public static $tracer = null;
+    /**
+     * @var array|null
+     */
+    public static $tracer;
 
-    public static $span = null;
+    /**
+     * @var OpenTracingSpan|null
+     */
+    public static $span;
 
-    public static $instance = null;
+    /**
+     * @var self|null
+     */
+    public static $instance;
 
     public static $disabled = false;
 
-    public static $propagator = \Jaeger\Constants\PROPAGATOR_JAEGER;
+    public static $propagator = PROPAGATOR_JAEGER;
 
-
-    private function __construct(){
-
-    }
-
-
-    private function __clone(){
-
-    }
-
-
-    public static function getInstance()
+    private function __construct()
     {
-        if(! (self::$instance instanceof self) )
-        {
+    }
+
+    private function __clone()
+    {
+    }
+
+    public static function getInstance(): ?Config
+    {
+        if (!(self::$instance instanceof self)) {
             self::$instance = new self();
         }
+
         return self::$instance;
     }
 
-
     /**
-     * init jaeger, return can use flush  buffers
-     * @param $serviceName
-     * @param string $agentHostPort
-     * @return Jaeger|null
-     * @throws \Exception
+     * init tracer.
+     *
+     * @throws RuntimeException
      */
-    public function initTracer($serverName, $agentHostPort = ''){
-
-        if(self::$disabled){
-            return NoopTracer::create();
+    public function initTracer(string $serviceName, string $agentHostPort = ''): Tracer
+    {
+        if (self::$disabled) {
+            return new NoopTracer();
         }
 
-        if($serverName == ''){
-            throw new \Exception("serverName require");
+        if ('' == $serviceName) {
+            throw new RuntimeException('serviceName require');
         }
 
-        if(isset(self::$tracer[$serverName]) && !empty(self::$tracer[$serverName])){
-            return self::$tracer[$serverName];
+        if (isset(self::$tracer[$serviceName]) && !empty(self::$tracer[$serviceName])) {
+            return self::$tracer[$serviceName];
         }
 
-
-        if($this->transport == null){
+        if (null == $this->transport) {
             $this->transport = new TransportUdp($agentHostPort);
         }
 
-        if($this->reporter == null) {
+        if (null == $this->reporter) {
             $this->reporter = new RemoteReporter($this->transport);
         }
 
-        if($this->sampler == null){
+        if (null == $this->sampler) {
             $this->sampler = new ConstSampler(true);
         }
 
-        if($this->scopeManager == null){
+        if (null == $this->scopeManager) {
             $this->scopeManager = new ScopeManager();
         }
 
-        $tracer = new Jaeger($serverName, $this->reporter, $this->sampler, $this->scopeManager);
+        $tracer = new Jaeger($serviceName, $this->reporter, $this->sampler, $this->scopeManager);
 
-        if($this->gen128bit == true){
+        if (true === $this->gen128bit) {
             $tracer->gen128bit();
         }
 
-        if(self::$propagator == \Jaeger\Constants\PROPAGATOR_ZIPKIN){
+        if (PROPAGATOR_ZIPKIN === self::$propagator) {
             $tracer->setPropagator(new ZipkinPropagator());
-        }else{
+        } else {
             $tracer->setPropagator(new JaegerPropagator());
         }
 
-
-        self::$tracer[$serverName] = $tracer;
-
+        self::$tracer[$serviceName] = $tracer;
 
         return $tracer;
     }
 
-
-    /**
-     * close tracer
-     * @param $disabled
-     */
-    public function setDisabled($disabled){
+    public function setDisabled(bool $disabled): Config
+    {
         self::$disabled = $disabled;
 
         return $this;
     }
 
-
-    public function setTransport(Transport\Transport $transport){
+    public function setTransport(Transport $transport): Config
+    {
         $this->transport = $transport;
 
         return $this;
     }
 
-
-    public function setReporter(Reporter $reporter){
+    public function setReporter(Reporter $reporter): Config
+    {
         $this->reporter = $reporter;
 
         return $this;
     }
 
-
-    public function setSampler(Sampler $sampler){
+    public function setSampler(Sampler $sampler): Config
+    {
         $this->sampler = $sampler;
 
         return $this;
     }
 
-
-    public function gen128bit(){
+    public function gen128bit(): Config
+    {
         $this->gen128bit = true;
 
         return $this;
     }
 
-
-    public function flush(){
-        if(count(self::$tracer) > 0) {
-            foreach(self::$tracer as $tracer){
+    public function flush(): bool
+    {
+        if (count(self::$tracer) > 0) {
+            foreach (self::$tracer as $tracer) {
                 $tracer->reportSpan();
             }
             $this->reporter->close();
