@@ -17,8 +17,8 @@ namespace Jaeger\Reporter;
 
 use Jaeger\Jaeger;
 
-interface Reporter{
-
+interface Reporter
+{
     public function report(Jaeger $jaeger);
 
     public function close();
