@@ -2,30 +2,36 @@
 
 namespace Jaeger;
 
+use OpenTracing\Scope as OpentracingScope;
+use OpenTracing\ScopeManager as OpentracingScopeManager;
+use OpenTracing\Span as OpentracingSpan;
 
-class ScopeManager implements \OpenTracing\ScopeManager{
-
+class ScopeManager implements OpentracingScopeManager
+{
     private $scopes = [];
 
-
     /**
-     * append scope
-     * @param \OpenTracing\Span $span
-     * @param bool $finishSpanOnClose
+     * {@inheritdoc}
+     *
+     * @param Span $span
+     *
      * @return Scope
      */
-    public function activate(\OpenTracing\Span $span, $finishSpanOnClose){
+    public function activate(OpentracingSpan $span, bool $finishSpanOnClose = self::DEFAULT_FINISH_SPAN_ON_CLOSE): OpentracingScope
+    {
         $scope = new Scope($this, $span, $finishSpanOnClose);
         $this->scopes[] = $scope;
+
         return $scope;
     }
 
-
     /**
-     * get last scope
-     * @return mixed|null
+     * {@inheritdoc}
+     *
+     * @return Scope
      */
-    public function getActive(){
+    public function getActive(): ?OpentracingScope
+    {
         if (empty($this->scopes)) {
             return null;
         }
@@ -33,20 +39,15 @@ class ScopeManager implements \OpenTracing\ScopeManager{
         return $this->scopes[count($this->scopes) - 1];
     }
 
-
-    /**
-     * del scope
-     * @param Scope $scope
-     * @return bool
-     */
-    public function delActive(Scope $scope){
+    public function deactivate(Scope $scope): bool
+    {
         $scopeLength = count($this->scopes);
 
-        if($scopeLength <= 0){
+        if ($scopeLength <= 0) {
             return false;
         }
 
-        for ($i = 0; $i < $scopeLength; $i++) {
+        for ($i = 0; $i < $scopeLength; ++$i) {
             if ($scope === $this->scopes[$i]) {
                 array_splice($this->scopes, $i, 1);
             }
