@@ -1,87 +1,140 @@
 <?php
-/*
- * Copyright (c) 2019, The Jaeger Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
- */
-
 namespace Jaeger\Thrift;
 
-use Thrift\Protocol\TProtocol;
+/**
+ * Autogenerated by Thrift Compiler (0.13.0)
+ *
+ * DO NOT EDIT UNLESS YOU ARE SURE THAT YOU KNOW WHAT YOU ARE DOING
+ *  @generated
+ */
+use Thrift\Base\TBase;
 use Thrift\Type\TType;
+use Thrift\Type\TMessageType;
+use Thrift\Exception\TException;
+use Thrift\Exception\TProtocolException;
+use Thrift\Protocol\TProtocol;
+use Thrift\Protocol\TBinaryProtocolAccelerated;
+use Thrift\Exception\TApplicationException;
 
-class Process implements TStruct{
+class Process
+{
+    static public $isValidate = false;
 
-    public static $tptl = null;
+    static public $_TSPEC = array(
+        1 => array(
+            'var' => 'serviceName',
+            'isRequired' => true,
+            'type' => TType::STRING,
+        ),
+        2 => array(
+            'var' => 'tags',
+            'isRequired' => false,
+            'type' => TType::LST,
+            'etype' => TType::STRUCT,
+            'elem' => array(
+                'type' => TType::STRUCT,
+                'class' => '\Jaeger\Thrift\Tag',
+                ),
+        ),
+    );
 
-    public static $serverName = '';
+    /**
+     * @var string
+     */
+    public $serviceName = null;
+    /**
+     * @var \Jaeger\Thrift\Tag[]
+     */
+    public $tags = null;
 
-    public static $thriftTags = [];
-
-    public static $wrote = '';
-
-    public function __construct($processThrift)
+    public function __construct($vals = null)
     {
-        self::$serverName = isset($processThrift['serverName']) ? $processThrift['serverName'] : '';
-        self::$thriftTags = isset($processThrift['tags']) ? $processThrift['tags'] : '';
-        self::$wrote = isset($processThrift['wrote']) ? $processThrift['wrote'] : '';
-    }
-
-
-    public function write(TProtocol $t){
-        self::$tptl = $t;
-
-        if(self::$wrote){
-            $tran = self::$tptl->getTransport();
-            $tran->write(self::$wrote);
-        } else {
-
-            self::$tptl->writeStructBegin("Process");
-
-            $this->handleProcessSName();
-            $this->handleProcessTags();
-            
-            self::$tptl->writeFieldStop();
-            self::$tptl->writeStructEnd();
-        }
-
-
-    }
-
-
-    private function handleProcessSName()
-    {
-        self::$tptl->writeFieldBegin("serverName", TType::STRING, 1);
-
-        self::$tptl->writeString(self::$serverName);
-
-        self::$tptl->writeFieldEnd();
-    }
-
-
-    private function handleProcessTags()
-    {
-        if(count(self::$thriftTags) > 0) {
-            self::$tptl->writeFieldBegin("tags", TType::LST, 2);
-            self::$tptl->writeListBegin(TType::STRUCT, count(self::$thriftTags));
-
-            $tagsObj = Tags::getInstance();
-            $tagsObj->setThriftTags(self::$thriftTags);
-            $tagsObj->write(self::$tptl);
-
-            self::$tptl->writeListEnd();
-            self::$tptl->writeFieldEnd();
+        if (is_array($vals)) {
+            if (isset($vals['serviceName'])) {
+                $this->serviceName = $vals['serviceName'];
+            }
+            if (isset($vals['tags'])) {
+                $this->tags = $vals['tags'];
+            }
         }
     }
 
+    public function getName()
+    {
+        return 'Process';
+    }
 
-    public function read(TProtocol $t){}
+
+    public function read($input)
+    {
+        $xfer = 0;
+        $fname = null;
+        $ftype = 0;
+        $fid = 0;
+        $xfer += $input->readStructBegin($fname);
+        while (true) {
+            $xfer += $input->readFieldBegin($fname, $ftype, $fid);
+            if ($ftype == TType::STOP) {
+                break;
+            }
+            switch ($fid) {
+                case 1:
+                    if ($ftype == TType::STRING) {
+                        $xfer += $input->readString($this->serviceName);
+                    } else {
+                        $xfer += $input->skip($ftype);
+                    }
+                    break;
+                case 2:
+                    if ($ftype == TType::LST) {
+                        $this->tags = array();
+                        $_size28 = 0;
+                        $_etype31 = 0;
+                        $xfer += $input->readListBegin($_etype31, $_size28);
+                        for ($_i32 = 0; $_i32 < $_size28; ++$_i32) {
+                            $elem33 = null;
+                            $elem33 = new \Jaeger\Thrift\Tag();
+                            $xfer += $elem33->read($input);
+                            $this->tags []= $elem33;
+                        }
+                        $xfer += $input->readListEnd();
+                    } else {
+                        $xfer += $input->skip($ftype);
+                    }
+                    break;
+                default:
+                    $xfer += $input->skip($ftype);
+                    break;
+            }
+            $xfer += $input->readFieldEnd();
+        }
+        $xfer += $input->readStructEnd();
+        return $xfer;
+    }
+
+    public function write($output)
+    {
+        $xfer = 0;
+        $xfer += $output->writeStructBegin('Process');
+        if ($this->serviceName !== null) {
+            $xfer += $output->writeFieldBegin('serviceName', TType::STRING, 1);
+            $xfer += $output->writeString($this->serviceName);
+            $xfer += $output->writeFieldEnd();
+        }
+        if ($this->tags !== null) {
+            if (!is_array($this->tags)) {
+                throw new TProtocolException('Bad type in structure.', TProtocolException::INVALID_DATA);
+            }
+            $xfer += $output->writeFieldBegin('tags', TType::LST, 2);
+            $output->writeListBegin(TType::STRUCT, count($this->tags));
+            foreach ($this->tags as $iter34) {
+                $xfer += $iter34->write($output);
+            }
+            $output->writeListEnd();
+            $xfer += $output->writeFieldEnd();
+        }
+        $xfer += $output->writeFieldStop();
+        $xfer += $output->writeStructEnd();
+        return $xfer;
+    }
 }
