@@ -1,223 +1,400 @@
 <?php
-/*
- * Copyright (c) 2019, The Jaeger Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
- */
-
 namespace Jaeger\Thrift;
 
-use Thrift\Protocol\TProtocol;
+/**
+ * Autogenerated by Thrift Compiler (0.13.0)
+ *
+ * DO NOT EDIT UNLESS YOU ARE SURE THAT YOU KNOW WHAT YOU ARE DOING
+ *  @generated
+ */
+use Thrift\Base\TBase;
 use Thrift\Type\TType;
+use Thrift\Type\TMessageType;
+use Thrift\Exception\TException;
+use Thrift\Exception\TProtocolException;
+use Thrift\Protocol\TProtocol;
+use Thrift\Protocol\TBinaryProtocolAccelerated;
+use Thrift\Exception\TApplicationException;
 
-class Span implements TStruct{
+class Span
+{
+    static public $isValidate = false;
 
-    public static $thriftSpan = null;
+    static public $_TSPEC = array(
+        1 => array(
+            'var' => 'traceIdLow',
+            'isRequired' => true,
+            'type' => TType::I64,
+        ),
+        2 => array(
+            'var' => 'traceIdHigh',
+            'isRequired' => true,
+            'type' => TType::I64,
+        ),
+        3 => array(
+            'var' => 'spanId',
+            'isRequired' => true,
+            'type' => TType::I64,
+        ),
+        4 => array(
+            'var' => 'parentSpanId',
+            'isRequired' => true,
+            'type' => TType::I64,
+        ),
+        5 => array(
+            'var' => 'operationName',
+            'isRequired' => true,
+            'type' => TType::STRING,
+        ),
+        6 => array(
+            'var' => 'references',
+            'isRequired' => false,
+            'type' => TType::LST,
+            'etype' => TType::STRUCT,
+            'elem' => array(
+                'type' => TType::STRUCT,
+                'class' => '\Jaeger\Thrift\SpanRef',
+                ),
+        ),
+        7 => array(
+            'var' => 'flags',
+            'isRequired' => true,
+            'type' => TType::I32,
+        ),
+        8 => array(
+            'var' => 'startTime',
+            'isRequired' => true,
+            'type' => TType::I64,
+        ),
+        9 => array(
+            'var' => 'duration',
+            'isRequired' => true,
+            'type' => TType::I64,
+        ),
+        10 => array(
+            'var' => 'tags',
+            'isRequired' => false,
+            'type' => TType::LST,
+            'etype' => TType::STRUCT,
+            'elem' => array(
+                'type' => TType::STRUCT,
+                'class' => '\Jaeger\Thrift\Tag',
+                ),
+        ),
+        11 => array(
+            'var' => 'logs',
+            'isRequired' => false,
+            'type' => TType::LST,
+            'etype' => TType::STRUCT,
+            'elem' => array(
+                'type' => TType::STRUCT,
+                'class' => '\Jaeger\Thrift\Log',
+                ),
+        ),
+    );
 
-    public static $tptl = null;
+    /**
+     * @var int
+     */
+    public $traceIdLow = null;
+    /**
+     * @var int
+     */
+    public $traceIdHigh = null;
+    /**
+     * @var int
+     */
+    public $spanId = null;
+    /**
+     * @var int
+     */
+    public $parentSpanId = null;
+    /**
+     * @var string
+     */
+    public $operationName = null;
+    /**
+     * @var \Jaeger\Thrift\SpanRef[]
+     */
+    public $references = null;
+    /**
+     * @var int
+     */
+    public $flags = null;
+    /**
+     * @var int
+     */
+    public $startTime = null;
+    /**
+     * @var int
+     */
+    public $duration = null;
+    /**
+     * @var \Jaeger\Thrift\Tag[]
+     */
+    public $tags = null;
+    /**
+     * @var \Jaeger\Thrift\Log[]
+     */
+    public $logs = null;
 
-    public static $instance = null;
-
-    private function __construct()
+    public function __construct($vals = null)
     {
-    }
-
-
-    private function __clone()
-    {
-    }
-
-
-    public static function getInstance(){
-        if(! (self::$instance instanceof self) )
-        {
-            self::$instance = new self();
+        if (is_array($vals)) {
+            if (isset($vals['traceIdLow'])) {
+                $this->traceIdLow = $vals['traceIdLow'];
+            }
+            if (isset($vals['traceIdHigh'])) {
+                $this->traceIdHigh = $vals['traceIdHigh'];
+            }
+            if (isset($vals['spanId'])) {
+                $this->spanId = $vals['spanId'];
+            }
+            if (isset($vals['parentSpanId'])) {
+                $this->parentSpanId = $vals['parentSpanId'];
+            }
+            if (isset($vals['operationName'])) {
+                $this->operationName = $vals['operationName'];
+            }
+            if (isset($vals['references'])) {
+                $this->references = $vals['references'];
+            }
+            if (isset($vals['flags'])) {
+                $this->flags = $vals['flags'];
+            }
+            if (isset($vals['startTime'])) {
+                $this->startTime = $vals['startTime'];
+            }
+            if (isset($vals['duration'])) {
+                $this->duration = $vals['duration'];
+            }
+            if (isset($vals['tags'])) {
+                $this->tags = $vals['tags'];
+            }
+            if (isset($vals['logs'])) {
+                $this->logs = $vals['logs'];
+            }
         }
-        return self::$instance;
     }
 
-
-    public function setThriftSpan($thriftSpan = []){
-        self::$thriftSpan = $thriftSpan;
-    }
-
-
-    public function getThriftSpan(){
-        return self::$thriftSpan;
-    }
-
-
-    public function write(TProtocol $t)
+    public function getName()
     {
-        self::$tptl = $t;
-        if(isset(self::$thriftSpan['wrote']) && self::$thriftSpan['wrote']){
-            $tran = self::$tptl->getTransport();
-            $tran->write(self::$thriftSpan['wrote']);
-        } else {
-            $this->handleSpan(self::$thriftSpan);
+        return 'Span';
+    }
+
+
+    public function read($input)
+    {
+        $xfer = 0;
+        $fname = null;
+        $ftype = 0;
+        $fid = 0;
+        $xfer += $input->readStructBegin($fname);
+        while (true) {
+            $xfer += $input->readFieldBegin($fname, $ftype, $fid);
+            if ($ftype == TType::STOP) {
+                break;
+            }
+            switch ($fid) {
+                case 1:
+                    if ($ftype == TType::I64) {
+                        $xfer += $input->readI64($this->traceIdLow);
+                    } else {
+                        $xfer += $input->skip($ftype);
+                    }
+                    break;
+                case 2:
+                    if ($ftype == TType::I64) {
+                        $xfer += $input->readI64($this->traceIdHigh);
+                    } else {
+                        $xfer += $input->skip($ftype);
+                    }
+                    break;
+                case 3:
+                    if ($ftype == TType::I64) {
+                        $xfer += $input->readI64($this->spanId);
+                    } else {
+                        $xfer += $input->skip($ftype);
+                    }
+                    break;
+                case 4:
+                    if ($ftype == TType::I64) {
+                        $xfer += $input->readI64($this->parentSpanId);
+                    } else {
+                        $xfer += $input->skip($ftype);
+                    }
+                    break;
+                case 5:
+                    if ($ftype == TType::STRING) {
+                        $xfer += $input->readString($this->operationName);
+                    } else {
+                        $xfer += $input->skip($ftype);
+                    }
+                    break;
+                case 6:
+                    if ($ftype == TType::LST) {
+                        $this->references = array();
+                        $_size7 = 0;
+                        $_etype10 = 0;
+                        $xfer += $input->readListBegin($_etype10, $_size7);
+                        for ($_i11 = 0; $_i11 < $_size7; ++$_i11) {
+                            $elem12 = null;
+                            $elem12 = new \Jaeger\Thrift\SpanRef();
+                            $xfer += $elem12->read($input);
+                            $this->references []= $elem12;
+                        }
+                        $xfer += $input->readListEnd();
+                    } else {
+                        $xfer += $input->skip($ftype);
+                    }
+                    break;
+                case 7:
+                    if ($ftype == TType::I32) {
+                        $xfer += $input->readI32($this->flags);
+                    } else {
+                        $xfer += $input->skip($ftype);
+                    }
+                    break;
+                case 8:
+                    if ($ftype == TType::I64) {
+                        $xfer += $input->readI64($this->startTime);
+                    } else {
+                        $xfer += $input->skip($ftype);
+                    }
+                    break;
+                case 9:
+                    if ($ftype == TType::I64) {
+                        $xfer += $input->readI64($this->duration);
+                    } else {
+                        $xfer += $input->skip($ftype);
+                    }
+                    break;
+                case 10:
+                    if ($ftype == TType::LST) {
+                        $this->tags = array();
+                        $_size13 = 0;
+                        $_etype16 = 0;
+                        $xfer += $input->readListBegin($_etype16, $_size13);
+                        for ($_i17 = 0; $_i17 < $_size13; ++$_i17) {
+                            $elem18 = null;
+                            $elem18 = new \Jaeger\Thrift\Tag();
+                            $xfer += $elem18->read($input);
+                            $this->tags []= $elem18;
+                        }
+                        $xfer += $input->readListEnd();
+                    } else {
+                        $xfer += $input->skip($ftype);
+                    }
+                    break;
+                case 11:
+                    if ($ftype == TType::LST) {
+                        $this->logs = array();
+                        $_size19 = 0;
+                        $_etype22 = 0;
+                        $xfer += $input->readListBegin($_etype22, $_size19);
+                        for ($_i23 = 0; $_i23 < $_size19; ++$_i23) {
+                            $elem24 = null;
+                            $elem24 = new \Jaeger\Thrift\Log();
+                            $xfer += $elem24->read($input);
+                            $this->logs []= $elem24;
+                        }
+                        $xfer += $input->readListEnd();
+                    } else {
+                        $xfer += $input->skip($ftype);
+                    }
+                    break;
+                default:
+                    $xfer += $input->skip($ftype);
+                    break;
+            }
+            $xfer += $input->readFieldEnd();
         }
+        $xfer += $input->readStructEnd();
+        return $xfer;
     }
 
-
-    public function read(TProtocol $t)
+    public function write($output)
     {
-        // TODO: Implement read() method.
-    }
-
-
-    private function handleSpan($span)
-    {
-        self::$tptl->writeStructBegin("Span");
-
-        self::$tptl->writeFieldBegin('traceIdLow', TType::I64, 1);
-        self::$tptl->writeI64($span['traceIdLow']);
-        self::$tptl->writeFieldEnd();
-
-        self::$tptl->writeFieldBegin('traceIdHigh', TType::I64, 2);
-        self::$tptl->writeI64($span['traceIdHigh']);
-        self::$tptl->writeFieldEnd();
-
-        self::$tptl->writeFieldBegin('spanId', TType::I64, 3);
-        self::$tptl->writeI64($span['spanId']);
-        self::$tptl->writeFieldEnd();
-
-        self::$tptl->writeFieldBegin('parentSpanId', TType::I64, 4);
-        self::$tptl->writeI64($span['parentSpanId']);
-        self::$tptl->writeFieldEnd();
-
-        self::$tptl->writeFieldBegin('operationName', TType::STRING, 5);
-        self::$tptl->writeString($span['operationName']);
-        self::$tptl->writeFieldEnd();
-
-        if (isset($span['references'])) {
-            $this->handleSpanRefes($span['references']);
+        $xfer = 0;
+        $xfer += $output->writeStructBegin('Span');
+        if ($this->traceIdLow !== null) {
+            $xfer += $output->writeFieldBegin('traceIdLow', TType::I64, 1);
+            $xfer += $output->writeI64($this->traceIdLow);
+            $xfer += $output->writeFieldEnd();
         }
-
-        self::$tptl->writeFieldBegin('flags', TType::I32, 7);
-        self::$tptl->writeI32($span['flags']);
-        self::$tptl->writeFieldEnd();
-
-        self::$tptl->writeFieldBegin('startTime', TType::I64, 8);
-        self::$tptl->writeI64($span['startTime']);
-        self::$tptl->writeFieldEnd();
-
-        self::$tptl->writeFieldBegin('duration', TType::I64, 9);
-        self::$tptl->writeI64($span['duration']);
-        self::$tptl->writeFieldEnd();
-
-        if (isset($span['tags'])) {
-            $this->handleSpanTags($span['tags']);
+        if ($this->traceIdHigh !== null) {
+            $xfer += $output->writeFieldBegin('traceIdHigh', TType::I64, 2);
+            $xfer += $output->writeI64($this->traceIdHigh);
+            $xfer += $output->writeFieldEnd();
         }
-
-        if (isset($span['logs'])) {
-            $this->handleSpanLogs($span['logs']);
+        if ($this->spanId !== null) {
+            $xfer += $output->writeFieldBegin('spanId', TType::I64, 3);
+            $xfer += $output->writeI64($this->spanId);
+            $xfer += $output->writeFieldEnd();
         }
-
-        self::$tptl->writeFieldStop();
-        self::$tptl->writeStructEnd();
-    }
-
-
-    private function handleSpanLogs($logs)
-    {
-        self::$tptl->writeFieldBegin('logs', TType::LST, 11);
-        self::$tptl->writeListBegin(TType::STRUCT, count($logs));
-
-        foreach ($logs as $log) {
-            $this->handleLog($log);
+        if ($this->parentSpanId !== null) {
+            $xfer += $output->writeFieldBegin('parentSpanId', TType::I64, 4);
+            $xfer += $output->writeI64($this->parentSpanId);
+            $xfer += $output->writeFieldEnd();
         }
-
-        self::$tptl->writeListEnd();
-        self::$tptl->writeFieldEnd();
-    }
-
-
-    private function handleLog($log)
-    {
-        self::$tptl->writeStructBegin("Log");
-
-        self::$tptl->writeFieldBegin('timestamp', TType::I64, 1);
-        self::$tptl->writeI64($log['timestamp']);
-        self::$tptl->writeFieldEnd();
-
-        $this->handleLogFields($log['fields']);
-
-        self::$tptl->writeFieldStop();
-        self::$tptl->writeStructEnd();
-    }
-
-
-    private function handleLogFields($fields)
-    {
-        self::$tptl->writeFieldBegin('fields', TType::LST, 2);
-        self::$tptl->writeListBegin(TType::STRUCT, count($fields));
-
-        $tagsObj = Tags::getInstance();
-        $tagsObj->setThriftTags($fields);
-        $tagsObj->write(self::$tptl);
-
-        self::$tptl->writeListEnd();
-        self::$tptl->writeFieldEnd();
-    }
-
-
-    private function handleSpanTags($tags)
-    {
-        self::$tptl->writeFieldBegin('tags', TType::LST, 10);
-        self::$tptl->writeListBegin(TType::STRUCT, count($tags));
-
-        $tagsObj = Tags::getInstance();
-        $tagsObj->setThriftTags($tags);
-        $tagsObj->write(self::$tptl);
-
-        self::$tptl->writeListEnd();
-        self::$tptl->writeFieldEnd();
-    }
-
-
-    private function handleSpanRefes($references)
-    {
-        self::$tptl->writeFieldBegin('references', TType::LST, 6);
-        self::$tptl->writeListBegin(TType::STRUCT, count($references));
-
-        foreach ($references as $refe) {
-            $this->handleSpanRefe($refe);
+        if ($this->operationName !== null) {
+            $xfer += $output->writeFieldBegin('operationName', TType::STRING, 5);
+            $xfer += $output->writeString($this->operationName);
+            $xfer += $output->writeFieldEnd();
         }
-
-        self::$tptl->writeListEnd();
-        self::$tptl->writeFieldEnd();
-    }
-
-
-    private function handleSpanRefe($refe)
-    {
-        self::$tptl->writeStructBegin("SpanRef");
-
-        self::$tptl->writeFieldBegin("refType", TType::I32, 1);
-        self::$tptl->writeI32($refe['refType']);
-        self::$tptl->writeFieldEnd();
-
-        self::$tptl->writeFieldBegin("traceIdLow", TType::I64, 2);
-        self::$tptl->writeI64($refe['traceIdLow']);
-        self::$tptl->writeFieldEnd();
-
-        self::$tptl->writeFieldBegin("traceIdHigh", TType::I64, 3);
-        self::$tptl->writeI64($refe['traceIdHigh']);
-        self::$tptl->writeFieldEnd();
-
-        self::$tptl->writeFieldBegin("spanId", TType::I64, 4);
-        self::$tptl->writeI64($refe['spanId']);
-        self::$tptl->writeFieldEnd();
-
-        self::$tptl->writeFieldStop();
-        self::$tptl->writeStructEnd();
+        if ($this->references !== null) {
+            if (!is_array($this->references)) {
+                throw new TProtocolException('Bad type in structure.', TProtocolException::INVALID_DATA);
+            }
+            $xfer += $output->writeFieldBegin('references', TType::LST, 6);
+            $output->writeListBegin(TType::STRUCT, count($this->references));
+            foreach ($this->references as $iter25) {
+                $xfer += $iter25->write($output);
+            }
+            $output->writeListEnd();
+            $xfer += $output->writeFieldEnd();
+        }
+        if ($this->flags !== null) {
+            $xfer += $output->writeFieldBegin('flags', TType::I32, 7);
+            $xfer += $output->writeI32($this->flags);
+            $xfer += $output->writeFieldEnd();
+        }
+        if ($this->startTime !== null) {
+            $xfer += $output->writeFieldBegin('startTime', TType::I64, 8);
+            $xfer += $output->writeI64($this->startTime);
+            $xfer += $output->writeFieldEnd();
+        }
+        if ($this->duration !== null) {
+            $xfer += $output->writeFieldBegin('duration', TType::I64, 9);
+            $xfer += $output->writeI64($this->duration);
+            $xfer += $output->writeFieldEnd();
+        }
+        if ($this->tags !== null) {
+            if (!is_array($this->tags)) {
+                throw new TProtocolException('Bad type in structure.', TProtocolException::INVALID_DATA);
+            }
+            $xfer += $output->writeFieldBegin('tags', TType::LST, 10);
+            $output->writeListBegin(TType::STRUCT, count($this->tags));
+            foreach ($this->tags as $iter26) {
+                $xfer += $iter26->write($output);
+            }
+            $output->writeListEnd();
+            $xfer += $output->writeFieldEnd();
+        }
+        if ($this->logs !== null) {
+            if (!is_array($this->logs)) {
+                throw new TProtocolException('Bad type in structure.', TProtocolException::INVALID_DATA);
+            }
+            $xfer += $output->writeFieldBegin('logs', TType::LST, 11);
+            $output->writeListBegin(TType::STRUCT, count($this->logs));
+            foreach ($this->logs as $iter27) {
+                $xfer += $iter27->write($output);
+            }
+            $output->writeListEnd();
+            $xfer += $output->writeFieldEnd();
+        }
+        $xfer += $output->writeFieldStop();
+        $xfer += $output->writeStructEnd();
+        return $xfer;
     }
 }
