@@ -17,8 +17,11 @@ namespace Jaeger\Transport;
 
 use Jaeger\Jaeger;
 
-interface Transport {
+interface Transport
+{
     public function append(Jaeger $jaeger);
 
     public function flush();
+
+    public function close();
 }
