@@ -13,23 +13,24 @@
  * the License.
  */
 
-namespace tests;
+namespace tests\Sampler;
 
+use Jaeger\Constants;
 use Jaeger\Sampler\ProbabilisticSampler;
 use PHPUnit\Framework\TestCase;
-use Jaeger\Constants;
 
 class ProbabilisticSamplerTest extends TestCase
 {
-
-    public function testProbabilisticSampler(){
+    public function testProbabilisticSampler(): void
+    {
         $sample = new ProbabilisticSampler(0.0001);
-        $this->assertTrue($sample->IsSampled() !== null);
+        static::assertNotNull($sample->IsSampled());
     }
 
-    public function testConstSamplerGetTag(){
+    public function testConstSamplerGetTag(): void
+    {
         $sample = new ProbabilisticSampler(0.0001);
         $tags = $sample->getTags();
-        $this->assertTrue($tags[Constants\SAMPLER_TYPE_TAG_KEY] == 'probabilistic');
+        static::assertEquals('probabilistic', $tags[Constants\SAMPLER_TYPE_TAG_KEY]);
     }
 }
